requirements.txt
djangorestframework==3.12.2
drf-complex-filter~=0.1
